DTSPB-1576: Address review comments
