+++ src/main/java/uk/gov/hmcts/probate/controller/LifeEventController.java
@@ -13,6 +13,7 @@ import uk.gov.hmcts.probate.model.ccd.raw.request.CallbackRequest;
 import uk.gov.hmcts.probate.model.ccd.raw.request.CaseDetails;
 import uk.gov.hmcts.probate.model.ccd.raw.response.CallbackResponse;
 import uk.gov.hmcts.probate.security.SecurityUtils;
+import uk.gov.hmcts.probate.service.LifeEventCCDService;
 import uk.gov.hmcts.probate.service.LifeEventService;
 import uk.gov.hmcts.probate.transformer.CallbackResponseTransformer;
 
@@ -25,13 +26,14 @@ import java.util.List;
 public class LifeEventController {
 
     private final CallbackResponseTransformer callbackResponseTransformer;
+    private final LifeEventCCDService lifeEventCCDService;
     private final LifeEventService lifeEventService;
     private final SecurityUtils securityUtils;
 
     @PostMapping(path = "/update")
     public ResponseEntity<CallbackResponse> update(@RequestBody CallbackRequest request) {
         final CaseDetails caseDetails = request.getCaseDetails();
-        lifeEventService.verifyDeathRecord(caseDetails, securityUtils.getSecurityDTO());
+        lifeEventCCDService.verifyDeathRecord(caseDetails, securityUtils.getSecurityDTO());
         return ResponseEntity.ok(callbackResponseTransformer.updateTaskList(request));
     }
