+++ src/main/java/uk/gov/hmcts/probate/service/DeathRecordService.java
@@ -52,24 +52,4 @@ public class DeathRecordService {
 
         return builder.build();
     }
-
-    @SuppressWarnings("squid:S2583")
-    public uk.gov.hmcts.probate.model.ccd.raw.DeathRecord mapDeathRecordCCD(V1Death v1Death) {
-        if (null == v1Death) {
-            return null;
-        }
-        final uk.gov.hmcts.probate.model.ccd.raw.DeathRecord.DeathRecordBuilder builder
-            = uk.gov.hmcts.probate.model.ccd.raw.DeathRecord.builder().systemNumber(v1Death.getId());
-        final Deceased deceased = v1Death.getDeceased();
-
-        if (nonNull(deceased)) {
-            builder.name(String.format("%s %s", deceased.getForenames(), deceased.getSurname()))
-                    .dateOfBirth(deceased.getDateOfBirth())
-                    .sex(null == deceased.getSex() ? null : deceased.getSex().getValue())
-                    .address(deceased.getAddress())
-                    .dateOfDeath(deceased.getDateOfDeath());
-        }
-
-        return builder.build();
-    }
 }
