+++ src/test/java/uk/gov/hmcts/probate/service/DeathRecordServiceTest.java
@@ -11,8 +11,6 @@ import java.util.Collections;
 import java.util.List;
 
 import static java.util.Arrays.asList;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class DeathRecordServiceTest {
@@ -86,50 +84,5 @@ public class DeathRecordServiceTest {
         final List<CollectionMember<DeathRecord>> deathRecordCollectionMembers = deathRecordService
                 .mapDeathRecords(asList(v1Death));
         assertEquals(deathRecordCollectionMembers.size(), 1);
-        final DeathRecord value = deathRecordCollectionMembers.get(0).getValue();
-    }
-    
-    @Test
-    public void shouldMapDeathRecordToCCDFormat() {
-        Deceased deceased = new Deceased();
-        deceased.setForenames("Firstname");
-        deceased.setSurname("LastName");
-        deceased.setSex(Deceased.SexEnum.INDETERMINATE);
-        deceased.setAddress("An address");
-        V1Death v1Death = new V1Death();
-        v1Death.setDeceased(deceased);
-        v1Death.setId(1234);
-
-        final uk.gov.hmcts.probate.model.ccd.raw.DeathRecord deathRecord =
-            deathRecordService.mapDeathRecordCCD(v1Death);
-
-
-        assertEquals(deathRecord.getSystemNumber(), v1Death.getId());
-        assertEquals(deathRecord.getName(), String.format("%s %s", deceased.getForenames(), deceased.getSurname()));
-        assertEquals(deathRecord.getDateOfBirth(), deceased.getDateOfBirth());
-        assertEquals(deathRecord.getDateOfDeath(), deceased.getDateOfDeath());
-        assertEquals(deathRecord.getSex(), deceased.getSex().getValue());
-        assertEquals(deathRecord.getAddress(), deceased.getAddress());
-    }
-    
-    @Test
-    public void shouldHandleNull() {
-        final uk.gov.hmcts.probate.model.ccd.raw.DeathRecord result =
-            deathRecordService.mapDeathRecordCCD(null);
-
-        assertNull(result);
-    }
-
-
-    @Test
-    public void shouldHandleNullDeceased() {
-        V1Death v1Death = new V1Death();
-        v1Death.setId(1234);
-
-        final uk.gov.hmcts.probate.model.ccd.raw.DeathRecord result =
-            deathRecordService.mapDeathRecordCCD(v1Death);
-
-        assertNotNull(null, result);
-        assertEquals(1234, result.getSystemNumber());
     }
 }
